+++ src/components/KModal/KModal.vue
@@ -358,7 +358,7 @@ onBeforeUnmount(async () => {
       font-weight: var(--kui-font-weight-regular, $kui-font-weight-regular);
       line-height: var(--kui-line-height-30, $kui-line-height-30);
       max-height: v-bind('maxHeightValue');
-      overflow: hidden;
+      overflow-y: auto;
     }
 
     .modal-header {